Fix local IP addr lookup
In cases where the default subnet for an IPv4 private network does not
match the locally configured subnet on the interface, the local IP lookup
would fail.

Example: local ip 10.0.0.100/24, but the default subnet for this private
range is 10.0.0.0/8.

It would also fail for any IPv6 lookups since no default mask is defined
for private IPv6 ranges.

This mehtod of checking if the target IP address lies within the subnet
configured on the interface fixes both of those problems by not reying
on the default subnet at all but by reading the configured subnet from
the interfaces.